changed double to int; snake now moves in int coordinates

+++ Snakes/Entity.java
@@ -1,9 +1,9 @@
 import java.awt.Graphics;
 
 public abstract class Entity{
-    protected double x,y;
+    protected int  x,y;
 
-    public Entity(double x, double y){
+    public Entity( int  x, int y){
       this.x = x;
       this.y = y;
     }
